Multilingual font rendering
* put your ttf file into "3rdparty\extras\xash-extras\gfx\fonts"
* Change "cl_font.c" file (FONT_COUNT, g_FontsPath, g_FontsSize) to add your ttf font

+++ engine/client/console.c
@@ -1087,7 +1087,7 @@ static void Field_Paste( field_t *edit )
 	// send as if typed, so insert / overstrike works properly
 	pasteLen = Q_strlen( cbd );
 	for( i = 0; i < pasteLen; i++ )
-		Field_CharEvent( edit, cbd[i] );
+		Field_CharEvent( edit, (byte)cbd[i] );
 }
 
 /*
